third_party/WebKit/LayoutTests/external/wpt/fullscreen/api/element-request-fullscreen-and-exit-iframe-manual-expected.txt

This is a testharness.js-based test.
FAIL Element#requestFullscreen() and Document#exitFullscreen() in iframe assert_equals: event target expected Element node <iframe allowfullscreen=""></iframe> but got Document node with 2 children
Harness: the test ran to completion.
